fix bug in is_in_subset
We did not check the value of bool note in the second
if-statement, resulting in the function returning true
if a value was noted in the subset, even when we did not
want that. This in turn affected is_allowed, so the user
could not enter a value if it was noted at another cell in
the same row, col or box.